refactor(ai-chat-components): split textarea controller to own file
prompt-line-controller.ts held both the PromptLineController contract (~70
lines) and the TextareaController implementation (~460 lines). Contributors
navigating to the interface had to scroll past 532 lines of textarea DOM
plumbing. The rich controller already sits in its own file; this brings
the textarea runtime in line with that pattern.

Three mechanical changes in one commit (they must land together — splitting
them would leave one of the two TYPING_TIMEOUT_MS declarations alive):

- Move TextareaController and its CSS-installation block into the new
  prompt-line-textarea-runtime.ts, mirroring prompt-line-rich-runtime.ts.
- Slim prompt-line-controller.ts to the contract only: SetContentUpdater,
  PromptLineControllerInit, PromptLineController.
- Consolidate TYPING_TIMEOUT_MS into prompt-line-constants.ts (already the
  shared home for PROMPT_LINE_MAX_BLOCK_SIZE and HISTORY_DEFAULTS); remove
  the independent copies from prompt-line-controller.ts and
  tiptap/typing-indicator.ts.

No behavior change. Public API surface is unchanged — index.ts re-exports
only the two types from prompt-line-controller.ts, which remain there.
